+++ src/components/controlled/transactions-table/components/index.ts
@@ -1,0 +1,7 @@
+export * from './TransactionAge';
+export * from './TransactionHash';
+export * from './TransactionIcon';
+export * from './TransactionMethod';
+export * from './TransactionValue';
+export * from './TransactionShards';
+export * from './TransactionAccount';