+++ src/components/controlled/transactions-table/transactions-table.tsx
@@ -3,6 +3,7 @@ import { Component, h, Prop } from '@stencil/core';
 import { DataTestIdsEnum } from '../../../constants/dataTestIds.enum';
 import type { TransactionRowType } from './transactions-table.type';
 import styles from './transactions-table.styles'
+import { TransactionAccount, TransactionAge, TransactionHash, TransactionMethod, TransactionShards, TransactionValue } from './components';
 
 const COLUMNS = ['Txn Hash', 'Age', 'Shard', 'From', 'To', 'Method', 'Value'];
 
@@ -30,45 +31,45 @@ export class TransactionsTable {
           {this.transactions.map(transaction => (
             <tr class={styles.transactionsTableBodyRow}>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-hash class={styles.transactionsTableBodyCellChild} transaction={transaction}></mvx-transaction-hash>
+                <TransactionHash class={styles.transactionsTableBodyCellChild} transaction={transaction} />
               </td>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-age
+                <TransactionAge
                   class={styles.transactionsTableBodyCellChild}
                   age={transaction.age.timeAgo}
                   tooltip={transaction.age.tooltip}
-                ></mvx-transaction-age>
+                />
               </td>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-shards class={styles.transactionsTableBodyCellChild} transaction={transaction}></mvx-transaction-shards>
+                <TransactionShards class={styles.transactionsTableBodyCellChild} transaction={transaction} />
               </td>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-account
+                <TransactionAccount
                   class={styles.transactionsTableBodyCellChild}
                   account={transaction.sender}
                   dataTestId={DataTestIdsEnum.transactionSender}
                   scope="sender"
                   showLockedAccounts={true}
-                ></mvx-transaction-account>
+                />
               </td>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-account
+                <TransactionAccount
                   class={styles.transactionsTableBodyCellChild}
                   account={transaction.receiver}
                   dataTestId={DataTestIdsEnum.transactionReceiver}
                   scope="receiver"
                   showLockedAccounts={true}
-                ></mvx-transaction-account>
+                />
               </td>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-method
+                <TransactionMethod
                   class={styles.transactionsTableBodyCellChild}
                   method={transaction.method.name}
                   actionDescription={transaction.method.actionDescription}
-                ></mvx-transaction-method>
+                />
               </td>
               <td class={styles.transactionsTableBodyCell}>
-                <mvx-transaction-value class={styles.transactionsTableBodyCellChild} value={transaction.value}></mvx-transaction-value>
+                <TransactionValue class={styles.transactionsTableBodyCellChild} value={transaction.value} />
               </td>
             </tr>
           ))}