+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Added the Storybook as a live server deployment](https://github.com/multiversx/mx-sdk-dapp-ui/pull/248)
 
+- [Refactor components in transactions table](https://github.com/multiversx/mx-sdk-dapp-ui/pull/243)
+
 ## [[0.0.35](https://github.com/multiversx/mx-sdk-dapp-ui/pull/246)] - 2025-10-10
 
 - [Added passkey provider](https://github.com/multiversx/mx-sdk-dapp-ui/pull/242)